+++ tests/QueryTest.php
@@ -627,12 +627,6 @@ class QueryTest extends MockeryTestCase
 
         $result = $query->first();
 
-        $this->assertInstanceOf(Person::class, $result);
-        $this->assertEquals(100, $result->id());
-        $this->assertEquals('Sherlock', $result->name);
-
-        $result = $query->first(1, true);
-
         $this->assertCount(1, $result);
         $this->assertInstanceOf(Person::class, $result[0]);
         $this->assertEquals(100, $result[0]->id());